requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
torch
torchaudio
transformers
huggingface_hub
sentence-transformers
faiss-cpu
pytesseract
Pillow
TTS
gradio
numpy
soundfile
